+++ src/config/webpack.config.js
@@ -1,0 +1,9 @@
+const path = require('path');
+
+module.exports = {
+  entry: './app.js',
+  output: {
+    path: path.resolve(__dirname, 'dist'),
+    filename: 'esperbook_bundle.js',
+  },
+};
