+++ src/repo/companyRepo.js
@@ -1,8 +1,9 @@
-var Company = require('../model/models');
+var Company = require('../model/company');
 const sql = require('mssql');
 require('dotenv').config();
 var dbConnection = require('../config/dbConnection'); 
 dbConnection.connect(dbConnection.config);
+const request = new sql.Request();
 
 function createCompany(req){
 return Company.Company.create(req).then(res=> {
@@ -18,8 +19,6 @@ console.log(companies);
 
 async function fetchCompany(companyToken,callback) {
   try {
-
-    const request = new sql.Request();
     request.input('param', sql.VarChar, companyToken);
     const query = process.env.QUERY_GET_COMPANY_DETAILS_BY_COMPANY_TOKEN;
     const result = await request.query(query);
